webapp/pyproject.toml
# Copyright 2024 Google LLC
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#     https://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

[project]
name = "buses-dashboard"
version = "0.1.0"
description = "This webapp will walk through some setup and guidelines on building your own data lakehouse, based on open source technologies to enable multi-platform data lakehouse architecture"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.1",
    "flask-executor>=1.0.0",
    "google-cloud-bigquery>=3.33.0",
    "google-cloud-core>=2.4.3",
    "confluent-kafka>=2.11.1",
    "google-cloud-logging>=3.12.1",
    "google-cloud-dataproc>=5.21.0",
    "google-cloud-storage>=2.19.0",
    "google-auth>=2.40.2",
]

[dependency-groups]
dev = [
    "black>=25.9.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
